Use Schema duration/byte tokens and ship an installable PWA
Schema is the settings SSOT for Go duration tokens and KiB/MiB/GiB.
Serve ungated ICO, manifest, and a fetch-handling service worker so the
dashboard is installable without caching live or inference traffic.

+++ internal/proxy/models.go
@@ -47,7 +47,7 @@ func (s *Server) serveModels(w http.ResponseWriter, r *http.Request, t *target, 
 	ctx, cancel := context.WithTimeout(r.Context(), cfg.ModelsDiscoveryTimeout)
 	defer cancel()
 	r = r.WithContext(ctx)
-	d := &modelsDiscovery{ctx: ctx, bytesLeft: cfg.ModelsDiscoveryMaxBytes, pagesLeft: cfg.ModelsDiscoveryMaxPages}
+	d := &modelsDiscovery{ctx: ctx, bytesLeft: int64(cfg.ModelsDiscoveryMaxBytes), pagesLeft: cfg.ModelsDiscoveryMaxPages}
 	key = s.resolveKeyTransparent(r, t, key)
 	switch t.format {
 	case "cursor":
